finished work with sqlite db
