+++ .gitignore
@@ -1,0 +1,3 @@
+# Default ignored files
+/__pycache__/
+/.idea/
