+++ main.py
@@ -1,18 +1,11 @@
 from flask import Flask, render_template, url_for, redirect, request
-from utility.helper import *
+from flask_sqlalchemy import SQLAlchemy
 
-l = ["0", "1", "2", "add", "font", "clone", "telugu", "english", "absolute"]
 app = Flask(__name__)
 app.config["SECRET_KEY"] = "5791628bb0b13ce0c676dfde280ba245"
-params = {"length": 5, "attempt": 6, "language": "English"}
-words = {
-    "solution": l[params["length"]],
-    "result": [],
-    "word_input": "",
-    "wordCount": 0,
-    "word_len_test": True,
-    "status": "PROCESS",
-}
+app.config["SQLALCHEMY_DATABASE_URI"] = "sqlite:///words.db"
+db = SQLAlchemy(app)
+from helper import *
 
 
 @app.route("/", methods=["GET", "POST"])
