+++ templates/base.html
@@ -106,7 +106,7 @@
                                                                 {{params.length}}
                                                             </option>
                                                             <div class="dropdown-menu ">
-                                                                {%for i in range(3,9)%}
+                                                                {%for i in range(3,13)%}
                                                                 <option class="dropdown-item " value="{{i}}">{{i}}</option> {%endfor%}
                                                             </div>
                                                         </select>
@@ -119,7 +119,7 @@
                                                                 {{params.attempt}}
                                                             </option>
                                                             <div class="dropdown-menu ">
-                                                                {%for i in range(5,13)%}
+                                                                {%for i in range(4,16)%}
                                                                 <option class="dropdown-item" value="{{i}}">{{i}}</option> {%endfor%}
                                                             </div>
                                                         </select>
